+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/infrastructure/plugins/DefaultClusterConfig.java
@@ -45,7 +45,7 @@ public final class DefaultClusterConfig {
     private static String standbyLogReplicationPort = "9020";
 
     @Getter
-    private static int logSenderBufferSize = 1;
+    private static int logSenderBufferSize = 2;
 
     @Getter
     private static int logSenderRetryCount = 5;
