+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/IllegalSnapshotEntrySizeException.java
@@ -1,0 +1,7 @@
+package org.corfudb.infrastructure.logreplication.replication.send;
+
+public class IllegalSnapshotEntrySizeException extends RuntimeException {
+    public IllegalSnapshotEntrySizeException(String message) {
+        super(message);
+    }
+}
