+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/IllegalTransactionStreamsException.java
@@ -6,3 +6,4 @@ public class IllegalTransactionStreamsException extends RuntimeException {
         super(message);
     }
 }
+
