+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/LogEntrySender.java
@@ -8,6 +8,7 @@ import org.corfudb.infrastructure.logreplication.replication.fsm.LogReplicationF
 import org.corfudb.infrastructure.logreplication.replication.fsm.LogReplicationEvent.LogReplicationEventType;
 import org.corfudb.infrastructure.logreplication.replication.send.logreader.LogEntryReader;
 import org.corfudb.infrastructure.logreplication.replication.send.logreader.ReadProcessor;
+import org.corfudb.infrastructure.logreplication.replication.send.logreader.StreamsLogEntryReader;
 import org.corfudb.protocols.wireprotocol.logreplication.LogReplicationEntry;
 import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.exceptions.TrimmedException;
@@ -32,7 +33,7 @@ public class LogEntrySender {
     /*
      * Implementation of Log Entry Reader. Default implementation reads at the stream layer.
      */
-    private LogEntryReader logEntryReader;
+    private StreamsLogEntryReader logEntryReader;
 
    /*
     * Implementation of buffering messages and sending/resending messages
@@ -67,7 +68,7 @@ public class LogEntrySender {
                           ReadProcessor readProcessor, LogReplicationFSM logReplicationFSM) {
 
         this.runtime = runtime;
-        this.logEntryReader = logEntryReader;
+        this.logEntryReader = (StreamsLogEntryReader)logEntryReader;
         this.dataSenderBufferManager = new LogEntrySenderBufferManager(dataSender);
         this.logReplicationFSM = logReplicationFSM;
     }
@@ -117,6 +118,11 @@ public class LogEntrySender {
                     // Request full sync (something is wrong I cant deliver)
                     // (Optimization):
                     // Back-off for couple of seconds and retry n times if not require full sync
+                }
+
+                if (logEntryReader.hasNoiseData()) {
+                    cancelLogEntrySync(LogReplicationError.ILLEGAL_TRANSACTION, LogReplicationEventType.REPLICATION_SHUTDOWN, logEntrySyncEventId);
+                    return;
                 }
 
             } catch (TrimmedException te) {
