+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/send/logreader/LogEntryReader.java
@@ -23,4 +23,6 @@ public interface LogEntryReader {
     void reset(long lastSentBaseSnapshotTimestamp, long lastAckedTimestamp);
 
     void setTopologyConfigId(long topologyConfigId);
+
+    boolean hasNoiseData();
 }
