+++ test/src/test/java/org/corfudb/integration/AbstractIT.java
@@ -564,7 +564,7 @@ public class AbstractIT extends AbstractCorfuTest {
          * @throws IOException
          */
         public Process runServer() throws IOException {
-            final String serverConsoleLogPath = "/Users/maxi/Projects/tmp/test.result"; //CORFU_LOG_PATH + File.separator + host + "_" + port + "_consolelog";
+            final String serverConsoleLogPath = CORFU_LOG_PATH + File.separator + host + "_" + port + "_consolelog";
 
             File logPath = new File(getCorfuServerLogPath(host, port));
             if (!logPath.exists()) {
