+++ test/src/test/java/org/corfudb/integration/ReplicationReaderWriterWithUFOIT.java
@@ -1,0 +1,4 @@
+package org.corfudb.integration;
+
+public class ReplicationReaderWriterWithUFOIT {
+}
