+++ test/src/test/resources/logback-test.xml
@@ -44,8 +44,14 @@
     </logger>
 
     <root level="INFO">
-        <appender-ref ref="FILE" />
-        <appender-ref ref="STDOUT" />
-        <appender-ref ref="MetricsRollingFile" />
+    <appender-ref ref="FILE" />
+    <appender-ref ref="STDOUT" />
+    <appender-ref ref="MetricsRollingFile" />
     </root>
+
+    <!--<root level="DEBUG">-->
+        <!--&lt;!&ndash;<appender-ref ref="FILE" />&ndash;&gt;-->
+        <!--<appender-ref ref="STDOUT" />-->
+        <!--&lt;!&ndash;<appender-ref ref="MetricsRollingFile" />&ndash;&gt;-->
+    <!--</root>-->
 </configuration>
